Move some PlatesRendererFactory tests to PlatesEngineFactory tests
